requirements.txt
astroid==2.4.2
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
idna==2.10
isort==5.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
limits==1.5.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
timeago==1.0.14
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
